Convert to Path::Tiny for all methods
This converts all methods over to using Path::Tiny for path
handling. This ensures consistency in path handling. In particular,
directories will always be represented the same way. If directory
names /tmp and /tmp/ are specified they both become /tmp which is very
useful when doing lookups in the registry.