Added support for specifying instance arguments in JSON.
This adds a new role which supports specifying the arguments for an
instance as a JSON string or file via a new_from_json method. This is
used in the ContentDecoder role and support is added to the associated
type coercion. This means it is possible to specify a class name and
it's arguments in a single string which suits a mapping from LCFG
resource values.