Improved content decoder list handling.
This alters the way the list of content decoders is split when it is
expressed as a single string. The split is now done on a | (pipe)
character with optional surrounding whitespace. This character should
hopefully not clash with anything that could appear in the JSON
specifications of the arguments for individual classes. Previously the
comma was also supported but that could cause problems.