Fixed bugs in symlink handling
Two bugs were found using the new tests:

The path_type_is_correct method was incorrectly checking if the path
was a file before checking if it was a symlink. This broke whenever
the symlink pointed to a directory. There was no need for the extra
check of file type so this has been simplified to just check if the
path is a symlink.

When an update to a symlink was required the build method was
attempting to delete the target file (which comes from the source
attribute) instead of the symlink. The log message was similarly
referring to the incorrect file.