Tweaked handling of changed files
This moves the logging of files changes from the FileBuilder to the
FileManager. This means that even when the FileManager is used in a
standalone manner the logging messages are generated.

The FileBuilder has also been altered to return a list of paths which
have been changed. This really needs some more thought about whether
it is necessary to differentiate between those files which are newly
created and those which were updated.