New Linux ACL manager
This adds a new AttributeManager interface which currently just
requires the existence of a configure method. Also added is a LinuxACL
manager which works in a similar way to the setfacl command. This is
really a proof of concept that the framework is useful. It still needs
to be plumbed into the FileManager role.