Added support for LCFG-style TT templates
This adds support for the LCFG-style Template Toolkit templates. The
new class inherits from the standard TT class and overrides the
default options to make it behave in the same way as LCFG::Template.

The handling of the default options was simplified so that it is
easier to override the method.