Added support for the old LCFG style templates
Frankly, it's a miracle this works. The code has been imported from
the LCFG::Template module and thoroughly cleaned and tidied. There are
no functionality changes other than the replacement of the
ReadTemplate method. Note that this consequently cannot handle reading
template text from stdin. That is really the job of the caller or it
might be possible to add a Xenon::Resource if really necessary.