Improved log handling
The logger will now be automatically initialised when necessary. If no
log configuration is specified then the default config file -
/etc/xenon/log.conf - will be used when it exists. If the default
config file does not exist then a basic screen logger will be created.

This change also adds a debug attribute which can be used to raise and
lower the log level when required.