Moved cleanup of unmanaged paths out of configure method
This moves the cleanup of unmanaged paths out of the configure method
into a separate cleanup_unmanaged_paths method. The logic has also
been reworked so that there are a lot fewer levels of braces which
makes it much more readable. This cleanup revealed a couple of small
bugs with the handling of the list of registered paths passed back
from the registry.