Added support for LCFG-style Base64 encoding
This adds support for the slightly bizarre support feature in the LCFG
file component for decoding Base64 where each line of data is encoded
separately and then the lines are concatenated with a *literal* \n
string (NOT a newline). An empty string between two literal \n strings
is interpreted as an empty line. Also added tests based on real-world
usage to ensure this works as expected. Note that this doesn't affect
the decoding of normal-style Base64 encoded data...