Reworked backup management
The management of backup files has been completely reworked. The
previous system was based on hard links which worked well when the
file was wholly managed by Xenon. Sadly it breaks badly when someone
edits the file directly since that also alters the hardlinked backup.

This new approach safely makes a temporary copy and then applies the
owner, group, mode, timestamps. If all that succeeds then any previous
version of the file is removed and the temporary file is renamed to
the required backup file name. This will not preserve ACLs, xattr,
etc. so that will have to be handled separately when necessary.

This change also adds a test suite for the backup system and
documentation of the API of the role.