Reworked the make_backup method
The make_backup method now returns a Path::Tiny object which
represents the backup file. If the style is "none" or the path is not
a file then undef is returned.

The code which generates the backup file name has now been split out
into a separate backup_file method to make it easier to modify.

Tests and documentation were modified to reflect the API changes.