Added new clobber option
This adds support for a "clobber" option which by default is set to
true. When it is set to false the "contents" of the file will not be
overwritten if it already exists. Any permissions, ACLs, attributes,
etc., will still be updated as normal though. For symlinks, setting this
to false will prevent the updating of the target for the link if it
already exists.