Reworked build method again
This splits out the code which decides which type of action is
required (i.e. create, update or none) into a separate change_type
method which makes it similar to the content_change_type method in the
FileContentManager role. The actually updating of the link is also
reworked to follow the build method in the FileContentManager in the
way it handles dryrun and clobber options. Definitely possibilities
for merging common behaviour at some point...