Improved SSL option handling
This improves the handling of SSL options for resources which are
accessible via https. When hostname verification is enabled the mode
is now explicitly set to SSL_VERIFY_PEER. The detection of standard
ca_path directories has been dropped in favour of using the default
(probably from Mozilla::CA) which is usually a lot more reliable.