Use same logger for all classes
This changes the way the logger is selected so that the same one is
used for all classes which consume the Log4perl role. Without this the
setting of log levels (e.g. debug) is not consistent across the
code. The expected behaviour is that enabling the debugging will turn
it on everywhere.