Individual file configurations now in separate processes
This uses the new fork_with_timeout functionality to configure
individual files in separate processes. As well as making it possible
to enforce a time limit this means it is not possible for the
configuration of a single file to mess up the environment of the file
builder process. This is particularly useful if the individual process
needs different environment variables to be set.